requirements.txt
flask==2.1.2
flask-restx>=0.5.1
pytest~=7.2.0
pytest-cov==4.0.0
flask-cors==3.0.10
werkzeug==2.0.2
requests
python-dotenv
pyjwt
cachetools
cryptography
hkdf
python-jose